Sanskrit_Backend/requirements.txt
flask==2.2.2
flask-pymongo==2.3.0
openai==0.27.7
python-dotenv==1.0.0
flask-login==0.6.1
bcrypt==4.0.1
requests==2.31.0